+++ docs/source/changes.md
@@ -9,6 +9,8 @@ releases are available on [PyPI](https://pypi.org/project/pytask) and
 
 - {pull}`633` adds support for Python 3.13 and drops support for 3.8.
 - {pull}`640` stops the live display when an exception happened during the execution.
+- {pull}`646` adds a `.gitignore` to the `.pytask/` folder to exclude it from version
+  control.
 
 ## 0.5.1 - 2024-07-20